extracted IPC logic into own script
