+++ test/integration/iteration.js
@@ -4,8 +4,8 @@ const assert = require("assert")
 const common = require("common-display-module")
 const simple = require("simple-mock")
 
+const ipc = require("../../src/ipc")
 const iteration = require("../../src/iteration")
-const storing = require("../../src/storing/storing")
 
 describe("Iteration", ()=>
 {
@@ -20,7 +20,7 @@ describe("Iteration", ()=>
     simple.mock(common, "getDisplaySettings").returnWith(settings)
 
     // Set up IPC connection
-    storing.init().then(() => done())
+    ipc.connect().then(() => done())
   })
 
   afterEach(()=>
@@ -50,7 +50,7 @@ describe("Iteration", ()=>
 
       // the BigQuery row entry, see design doc for individual element description
       const row = data.data
-      assert.equal(row.event, "info")
+      assert.equal(row.event, "data")
       assert.equal(row.event_details, "")
       assert.equal(row.display_id, "DIS123")
       // ts will be inserted in logging module, so we won't be checking it here
