+++ test/unit/storing/storing.js
@@ -3,6 +3,8 @@
 const assert = require("assert")
 const common = require("common-display-module")
 const simple = require("simple-mock")
+
+const ipc = require("../../../src/ipc")
 const storing = require("../../../src/storing/storing")
 
 describe("Storing", ()=>
@@ -17,8 +19,7 @@ describe("Storing", ()=>
     simple.mock(common, "connect").returnWith(connection)
     simple.mock(common, "getDisplaySettings").returnWith(settings)
 
-    // Set up IPC connection
-    storing.init().then(() => done())
+    ipc.connect().then(() => done())
   })
 
   afterEach(()=>
@@ -53,7 +54,7 @@ describe("Storing", ()=>
 
     // the BigQuery row entry, see design doc for individual element description
     const row = data.data
-    assert.equal(row.event, "info")
+    assert.equal(row.event, "data")
     assert.equal(row.event_details, "")
     assert.equal(row.display_id, "DIS123")
     // ts will be inserted in logging module, so we won't be checking it here
